sdk-python/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "primitivedotdev"
version = "0.1.1"
description = "Official Primitive Python SDK for webhook handling and API access"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Primitive", email = "support@primitive.dev" }]
keywords = ["primitive", "email", "webhook", "signature", "hmac"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Typing :: Typed",
]
dependencies = [
  "attrs>=23.2.0,<26",
  "httpx>=0.27.0,<1",
  "jsonschema>=4.24.0,<5",
  "pydantic>=2.11,<3",
  "python-dateutil>=2.9.0,<3",
  "typing-extensions>=4.12.0; python_version < '3.11'",
]

[dependency-groups]
dev = [
  "basedpyright>=1.31.0,<2",
  "build>=1.2.2,<2",
  "datamodel-code-generator>=0.31.0,<1",
  "hatchling>=1.27.0,<2",
  "openapi-python-client>=0.27.0,<1",
  "pytest>=8.3.0,<9",
  "pytest-cov>=6.2.0,<7",
  "ruff>=0.12.0,<1",
  "twine>=6.1.0,<7",
]

[project.urls]
Homepage = "https://primitive.dev"
Repository = "https://github.com/primitivedotdev/sdks"
Issues = "https://github.com/primitivedotdev/sdks/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/primitive"]

[tool.hatch.build.targets.wheel.force-include]
"src/primitive/schemas/email_received_event.schema.json" = "primitive/schemas/email_received_event.schema.json"
"src/primitive/py.typed" = "primitive/py.typed"

[tool.hatch.build.targets.sdist]
include = [
  "/README.md",
  "/pyproject.toml",
  "/scripts",
  "/src/primitive",
  "/tests",
]

[tool.hatch.build.targets.sdist.force-include]
"../json-schema" = "json-schema"
"../test-fixtures" = "test-fixtures"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["primitive"]
omit = [
  "src/primitive/api/api/*",
  "src/primitive/api/client.py",
  "src/primitive/api/errors.py",
  "src/primitive/api/models/*",
  "src/primitive/api/types.py",
  "src/primitive/models_generated.py",
  "src/primitive/schemas/*",
]

[tool.coverage.report]
omit = [
  "src/primitive/api/api/*",
  "src/primitive/api/client.py",
  "src/primitive/api/errors.py",
  "src/primitive/api/models/*",
  "src/primitive/api/types.py",
  "src/primitive/models_generated.py",
  "src/primitive/schemas/*",
]

[tool.ruff]
target-version = "py310"
line-length = 88
extend-exclude = [
  "src/primitive/api/api",
  "src/primitive/api/client.py",
  "src/primitive/api/errors.py",
  "src/primitive/api/models",
  "src/primitive/api/types.py",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "SIM108", "UP012", "UP017"]

[tool.basedpyright]
pythonVersion = "3.10"
typeCheckingMode = "standard"
include = ["src", "tests", "scripts"]
exclude = [
  "src/primitive/api/api",
  "src/primitive/api/client.py",
  "src/primitive/api/errors.py",
  "src/primitive/api/models",
  "src/primitive/api/types.py",
  "src/primitive/models_generated.py",
  "src/primitive/schemas",
]
